+++ lib/ratel.dart
@@ -7,6 +7,7 @@ export 'core/logger.dart';
 export 'core/request_context.dart';
 export 'core/middleware.dart';
 export 'core/router.dart';
+export 'core/openapi.dart';
 export 'database/database.dart';
 export 'database/repository.dart';
 export 'dependency_injector/binding.dart';